Prevent from over-writing .npy files already present

+++ pre-post-proc/NRRD_NPY_Conv.py
@@ -26,8 +26,11 @@ if arguments.review == 'y':
 else:
     review_flag = False
 
-IMG_PATH = f"{FILE_PATH}SparseVolImgs/"
-SEG_PATH = f"{FILE_PATH}SparseVolMasks/"
+# IMG_PATH = f"{FILE_PATH}SparseVolImgs/"
+# SEG_PATH = f"{FILE_PATH}SparseVolMasks/"
+IMG_PATH = f"{FILE_PATH}TestVolImages/"
+SEG_PATH = f"{FILE_PATH}TestVolMasks/"
+print(IMG_PATH)
 subject_list = os.listdir(SEG_PATH)
 subject_list.sort()
 
@@ -45,30 +48,31 @@ for subject in subject_list:
         seg_vol = seg_vol[0, ...]
 
         img = seg[:-10] + "L.nrrd"
-    
-        try:
-            img_vol, _ = nrrd.read(f"{subject_img_path}/{img}")
-        except:
-            print(f"NO CORRESPONDING MASK FOUND: {seg} {img}")
-        else:
-            img_vol = (img_vol - min_val) / (max_val - min_val)
 
-            if img_vol.shape != seg_vol.shape:
-                print(f"{img} {seg} INCORRECT DIMENSION")
+        if not os.path.exists(f"{SAVE_PATH}Img/{img[:-5]}.npy"):
+            try:
+                img_vol, _ = nrrd.read(f"{subject_img_path}/{img}")
+            except:
+                print(f"NO CORRESPONDING MASK FOUND: {seg} {img}")
+            else:
+                img_vol = (img_vol - min_val) / (max_val - min_val)
 
-            if review_flag:
-                for i in range(seg_vol.shape[2]):
-                    fig, axs = plt.subplots(1, 3)                   
-                    axs[0].imshow(np.fliplr(img_vol[:, :, i].T), vmin=0.12, vmax=0.18, cmap='gray', origin='lower')
-                    axs[1].imshow(np.fliplr(seg_vol[:, :, i].T), vmin=0.12, vmax=0.18, cmap='gray', origin='lower')
-                    axs[2].imshow(np.fliplr(img_vol[:, :, i].T * seg_vol[:, :, i].T), vmin=0.12, vmax=0.18, cmap='gray', origin='lower')
+                if img_vol.shape != seg_vol.shape:
+                    print(f"{img} {seg} INCORRECT DIMENSION: {img_vol.shape} {seg_vol.shape}")
 
-                    plt.show()
-                    plt.close()
-            
-            np.save(f"{SAVE_PATH}Img/{img[:-5]}.npy", img_vol)
-            np.save(f"{SAVE_PATH}Seg/{seg[:-9]}.npy", seg_vol)
-            print(f"{img[:-5]}.npy, {seg[:-9]}.npy CONVERTED")
+                if review_flag:
+                    for i in range(seg_vol.shape[2]):
+                        fig, axs = plt.subplots(1, 3)
+                        axs[0].imshow(np.fliplr(img_vol[:, :, i].T), vmin=0.12, vmax=0.18, cmap='gray', origin='lower')
+                        axs[1].imshow(np.fliplr(seg_vol[:, :, i].T), vmin=0.12, vmax=0.18, cmap='gray', origin='lower')
+                        axs[2].imshow(np.fliplr(img_vol[:, :, i].T * seg_vol[:, :, i].T), vmin=0.12, vmax=0.18, cmap='gray', origin='lower')
+
+                        plt.show()
+                        plt.close()
+
+                np.save(f"{SAVE_PATH}Img/{img[:-5]}.npy", img_vol)
+                np.save(f"{SAVE_PATH}Seg/{seg[:-9]}.npy", seg_vol)
+                print(f"{img[:-5]}.npy, {seg[:-9]}.npy CONVERTED")
 
 if len(os.listdir(f"{SAVE_PATH}Img/")) != len(os.listdir(f"{SAVE_PATH}Seg/")):
     raise ValueError("Unequal number of converted vols")
